refactor: extract xrm builder panel overlays styles and types
