+++ storybook/src/form/xrm-form/form-xml-builder-panel/index.ts
@@ -1,0 +1,1 @@
+export * from './FormXmlBuilderPanel'
