+++ storybook/src/stories/form/XrmFormXmlBuilderStory.tsx
@@ -4,7 +4,7 @@ import { parseFormXml } from '@talxis/client-metadata'
 import { XrmForm } from '@talxis/base-controls/components/Form'
 import type { IXrmFormStrategy } from '@talxis/base-controls/components/Form'
 import { MemoryStrategy } from '@talxis/base-controls/components/Form'
-import { FormXmlBuilderPanel } from '../../form/xrm-form/FormXmlBuilderPanel'
+import { FormXmlBuilderPanel } from '../../form/xrm-form/form-xml-builder-panel'
 import { FormXmlEditor } from '../../form/xrm-form/FormXmlEditor'
 import { defaultFormXml } from '../../form/xrm-form/defaultFormXml'
 import { createModelStore } from '../../form/shared/modelStore'
